requirements.txt
flask>=3.0
waitress>=3.0
pdfplumber>=0.11
pypdfium2>=4.30
rapidocr-onnxruntime>=1.3
Pillow>=10.0
reportlab>=4.0
werkzeug>=3.0
